requirement.txt
opencv-python
numpy
pillow
google-cloud-vision
python-dotenv
tflite-runtime
